panties-server/requirements.txt
Django>=4.2,<5.0
django-allauth>=0.57.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
gunicorn>=21.2.0
dj-database-url>=2.1.0
